restructure nvim filesystem, cleanup plugins
