--- nvim/lua/custom/plugins/harpoon.lua
@@ -1,2 +1,0 @@
-return {}
-
