--- nvim/lua/custom/plugins/project.lua
@@ -1,6 +1,0 @@
-require('project_nvim').setup {
-  detection_methods = { 'pattern' },
-  patterns = { '.git', '.project', 'your_custom_marker_here' }
-}
-
-return {}
