--- nvim/lua/custom/plugins/telescope.lua
@@ -1,19 +1,0 @@
-require'telescope'.load_extension('projects')
-require('telescope').setup {
-  defaults = {
-    mappings = {
-      i = {
-        ['<C-u>'] = false,
-        ['<C-d>'] = false,
-      },
-    },
-  },
-  pickers = {
-    find_files = {
-      hidden = true
-    }
-  }
-}
-
-return {}
-
