+++ nvim/lua/kevincardona/init.lua
@@ -1,0 +1,4 @@
+require("kevincardona.lazy")
+require("kevincardona.set")
+require("kevincardona.remap")
+
