+++ nvim/lua/kevincardona/plugins/trouble.lua
@@ -1,0 +1,4 @@
+require('trouble').setup {
+    icons = false
+}
+return {}
